my_capital_bot/requirements.txt
python-dotenv
requests
pandas
numpy
scikit-learn
torch
